plugins(fixfileExtensions): organize
