+++ src/equicordplugins/fixFileExtensions/components.tsx
@@ -1,0 +1,12 @@
+const extensionMap = {
+    "ogg": [".ogv", ".oga", ".ogx", ".ogm", ".spx", ".opus"],
+    "jpg": [".jpg", ".jpeg", ".jfif", ".jpe", ".jif", ".jfi", ".pjpeg", ".pjp"],
+    "svg": [".svgz"],
+    "mp4": [".m4v", ".m4a", ".m4r", ".m4b", ".m4p"],
+    "mov": [".movie", ".qt"],
+};
+
+export const reverseExtensionMap = Object.entries(extensionMap).reduce((acc, [target, exts]) => {
+    exts.forEach(ext => acc[ext] = `.${target}`);
+    return acc;
+}, {} as Record<string, string>);
